+++ Reach-Coder/src/utopia/coder/reach/controller/app/ReachCoderAppLogic.scala
@@ -8,7 +8,7 @@ import utopia.coder.reach.controller.writer.ComponentFactoryWriter
 import utopia.coder.reach.util.Common
 import utopia.flow.parse.file.FileExtensions._
 import utopia.flow.time.Today
-import utopia.flow.util.TryExtensions._
+import utopia.flow.util.result.TryExtensions._
 import utopia.flow.util.console.CommandArguments
 import utopia.flow.util.logging.Logger
 import utopia.flow.view.immutable.View