+++ Vault-Coder/Changes.md
@@ -7,6 +7,7 @@ Compatible with Vault v2.1
 ### Other changes
 - Targeting is now the default mode
   - Replaced the `T` flag with `NT` (no-targeting)
+- When combo classes are present, the stored class version is now a trait, extended by said combined models
 - Generated Filter-traits now extend **Filterable** instead of **FilterableView**
 - Value-to-enumeration conversion is now more reliable for Int-based enumerations
 - Generated Int-based in-conditions now use `Condition.indexIn(...)`